fix: Apply CodeRabbit critical fixes for reliability and security
Critical fixes from CodeRabbit review:
- Fix context key inconsistency (handler_test.go) - use 'jwt_claims' key
- Fix io.EOF handling in stream closure (grpc/client_test.go) - treat as clean termination
- Fix jwt.SignedString error handling (auth_test.go) - check and fail on errors
- Fix None checks for max_tokens/temperature (llm_service.py) - preserve zero values
- Replace deprecated typing imports (llm_service.py) - use AsyncIterator from collections.abc
- Alphabetize environment variables (.env.example) - fix dotenv-linter expectations
- Remove port exposure for production (docker-compose.selfhosted.yml) - postgres and redis
- Fix gRPC healthcheck (docker-compose.selfhosted.yml) - add real probe on port 50051
- Pin MinIO version (docker-compose.selfhosted.yml) - use RELEASE.2024-01-16T23-15-27Z
- Remove version key (docker-compose.selfhosted.yml) - use Compose v2+ format
- Fix nginx timeouts (nginx.conf) - reduce from 7d to 1h for WebSocket resources
- Remove deprecated X-XSS-Protection header (nginx.conf) - replace with Content-Security-Policy
- Change INTEGER to BIGINT (01-schema.sql) - support files >2GB
- Add gzip_types for common MIME types (nginx.conf)

Security improvements:
- Better error handling prevents invalid tokens
- Reduced WebSocket timeouts prevent resource exhaustion
- Modern CSP header provides better XSS protection
- Production deployment no longer exposes internal ports

+++ infra/docker/init-scripts/01-schema.sql
@@ -44,7 +44,7 @@ CREATE TABLE IF NOT EXISTS attachments (
     filename VARCHAR(255) NOT NULL,
     mime_type VARCHAR(100) NOT NULL,
     storage_path VARCHAR(500) NOT NULL,
-    size_bytes INTEGER,
+    size_bytes BIGINT,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW()
 );
